# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: LESS API
  version: 0.1.0
  description: |
    LESSのREST API定義。本fileがフロントエンドとバックエンドの契約の正本である。

    現在のスコープは以下とする。
      - Phase 0: 認証 (F-001)
      - Phase 1: Category初期data / Item CRUD / 一覧filter / Tag / 使用記録 (F-003〜F-006)
      - Phase 2: StorageUnit / 収納割当 / 容量集計 / 携行区分 (F-008 / F-009)

    Review (Phase 3)、PurchaseCandidate (Phase 4)、Scenario (Phase 5)、
    Import/Export (Phase 6) はまだ定義しない。

    Phase 2で設計書と異なる判断をした点:
      - 更新は 設計書 12.4 の `PUT /api/storage-units/{publicId}` に従い全置換とする。
        PATCHによる部分更新は導入しない (Itemの `updateItem` と同じ方針)。
      - 重量・容積のfield名は設計書 13.7 / 13.8 のcolumn名に合わせ単数形
        (`tareWeightGram` / `maximumVolumeMilliliter` 等) とする。
        Phase 1の `weightGram` / `volumeMilliliter` と揃える。
      - 設計書 12.4 に無い `restore` / `contents` / 割当の個別操作 /
        `GET /items/{publicId}/storage-allocations` を追加した。
        いずれもPhase 2の要件 (収納内容編集・分割割当・未割当数量) に必要である。
      - 収納単位の親集計は設計書 16.2 が子孫を規定していないため、
        直接割当分 (`itemWeightGram`) と子孫分 (`descendantWeightGram`) を
        別fieldで返し、`totalWeightGram` で合算する。二重計上を避けるため
        子孫分には子孫自身の自重を含め、親の自重は含めない。

    共通仕様 (設計書 12.1):
      - base path: `/api`
      - Content-Type: `application/json`
      - datetime: RFC3339 (UTC)
      - JSON field: camelCase
      - enum value: lowercase snake_case
      - path: 複数形kebab-case
      - URLでは内部IDではなくpublicIdを使用する

    一覧のpagination:
      設計書 12.2 はcursor paginationを基本とするが、Item一覧は
      任意のsort keyとの組み合わせ、および総件数表示を要件とするため
...

...

...
servers:
...
paths:
...
  /items/{publicId}/storage-allocations:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [items]
      operationId: listItemStorageAllocations
      summary: 所持品の収納割当を取得する
      description: |
        1つのアイテムがどの収納単位へ何個ずつ入っているかを返す。
        同一アイテムを複数収納単位へ分割割当できるため複数件になる (F-009)。

        1アイテムの割当件数は収納単位数で上限が決まり少数であるため、
        paginationを行わず全件を返す。
        `unassignedQuantity` はDBへ保存せず取得時に算出する。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemStorageAllocationListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units:
    get:
      tags: [storage-units]
      operationId: listStorageUnits
      summary: 収納単位一覧を取得する
      description: |
        認証ユーザーの収納単位を検索・絞り込み・並び替えして返す (F-008)。
        archive済みは既定で除外し、`includeArchived=true` の場合のみ含める。

        各要素は `capacity` に重量・容積の集計と超過判定を含む。
        集計は子孫収納単位を含むため、pageに含まれない収納単位も
        server側で階層をたどって計算する。
      security:
        - sessionCookie: []
      parameters:
        - name: keyword
          in: query
          description: 収納単位名または説明の部分一致 (大文字小文字を区別しない)
          schema:
            type: string
            maxLength: 100
        - name: storageTypeCode
          in: query
          schema:
            $ref: "#/components/schemas/StorageTypeCode"
        - name: mobilityClassCode
          in: query
          schema:
            $ref: "#/components/schemas/MobilityClassCode"
        - name: parentStorageUnitPublicId
          in: query
          description: 指定した収納単位の直接の子だけを返す。
          schema:
            type: string
            format: uuid
        - name: rootOnly
          in: query
          description: |
            trueの場合、親を持たない収納単位だけを返す。
            `parentStorageUnitPublicId` と同時に指定した場合は 400 を返す。
          schema:
            type: boolean
        - name: includeArchived
          in: query
          description: |
            archive済みの収納単位を含めるか。未指定時はfalse。
            Itemの `includeDeleted` と役割は同じだが、収納単位はarchiveのみを
            利用者へ見せるため名称を `includeArchived` とした。
          schema:
            type: boolean
        - name: sort
          in: query
          description: 並び替えkey。未指定時は sortOrder。
          schema:
            $ref: "#/components/schemas/StorageUnitSortKey"
        - name: order
          in: query
          description: 並び順。未指定時は asc (sortOrder) / desc (updatedAt) ではなく常に指定値を使う。未指定時は asc。
          schema:
            $ref: "#/components/schemas/SortOrder"
        - $ref: "#/components/parameters/LimitQueryParameter"
        - $ref: "#/components/parameters/OffsetQueryParameter"
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [storage-units]
      operationId: createStorageUnit
      summary: 収納単位を登録する
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageUnitRequest"
      responses:
        "201":
          description: 登録に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: |
            指定した親収納単位が存在しない。
            他ユーザーの収納単位を指定した場合も本statusを返す (設計書 18.3)。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: |
            業務ルール違反。以下を含む。
              - 階層が3を超える (`STORAGE_HIERARCHY_TOO_DEEP`)
              - archive済みの収納単位を親に指定した (`STORAGE_UNIT_PARENT_ARCHIVED`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [storage-units]
      operationId: getStorageUnitByPublicId
      summary: 収納単位を取得する
      description: |
        他ユーザーのpublicIdを指定した場合も 404 を返し、
        存在有無を公開しない (設計書 18.3)。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [storage-units]
      operationId: updateStorageUnit
      summary: 収納単位を更新する
      description: |
        全項目を置き換える。省略した任意項目はNULLへ更新される。
        `expectedVersion` が現在versionと一致しない場合は 409 を返す。

        `parentStorageUnitPublicId` の変更で階層が移動する。
        自分自身・子孫を親に指定した場合、および移動後の階層が3を超える場合は
        422 を返す。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStorageUnitRequest"
      responses:
        "200":
          description: 更新に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: |
            業務ルール違反。以下を含む。
              - 自分自身を親に指定した (`STORAGE_UNIT_SELF_PARENT`)
              - 子孫を親に指定した (`STORAGE_UNIT_CIRCULAR_PARENT`)
              - 階層が3を超える (`STORAGE_HIERARCHY_TOO_DEEP`)
              - archive済みの収納単位を親に指定した (`STORAGE_UNIT_PARENT_ARCHIVED`)
              - archive済みの収納単位を編集した (`STORAGE_UNIT_ARCHIVED`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}/archive:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    post:
      tags: [storage-units]
      operationId: archiveStorageUnit
      summary: 収納単位をarchiveする
      description: |
        soft deleteとして `archivedAt` を設定する (設計書 1.4)。

        以下の場合は 422 を返し、親のarchiveで子を暗黙にarchiveしない。
          - archive前の子収納単位が残っている (`STORAGE_UNIT_HAS_CHILDREN`)
          - 収納割当が残っている (`STORAGE_UNIT_HAS_ALLOCATIONS`)

        いずれも利用者へ「子を先にarchiveする」「中身を先に出す」ことを求める。
        暗黙のcascadeを行わないことで、復元時の状態を予測可能にする。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageUnitVersionRequest"
      responses:
        "200":
          description: archiveに成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}/restore:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    post:
      tags: [storage-units]
      operationId: restoreStorageUnit
      summary: archive済みの収納単位を復元する
      description: |
        親がarchive済みの場合は 422 を返す (`STORAGE_UNIT_PARENT_ARCHIVED`)。
        階層の不整合を作らないため、親から順に復元させる。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageUnitVersionRequest"
      responses:
        "200":
          description: 復元に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}/capacity:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [storage-units]
      operationId: getStorageUnitCapacity
      summary: 収納単位の重量・容積を取得する
      description: |
        設計書 16.2 の使用量と 16.3 の超過警告を返す。

        収納済みアイテムの一部に重量または容積が未設定の場合、
        既知分だけを合計し `hasUnknownWeight` / `hasUnknownVolume` を trueとする。
        未設定値を0として完全な値に見せない。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitCapacityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}/contents:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    get:
      tags: [storage-units]
      operationId: getStorageUnitContents
      summary: 収納単位の内容を取得する
      description: |
        直接割当されているアイテム、直接の子収納単位、容量集計をまとめて返す。
        収納内容編集画面が1requestで必要な情報を得られるようにする。

        1収納単位の割当件数は実運用で少数に収まるためpaginationを行わない。
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 取得に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitContentsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /storage-units/{publicId}/allocations:
    parameters:
      - $ref: "#/components/parameters/PublicIdPathParameter"

    post:
      tags: [storage-units]
      operationId: createStorageAllocation
      summary: 所持品を収納単位へ割り当てる
      description: |
        数量付きで割り当てる (F-009)。

        制約:
          - 割当数量は1以上
          - 同一収納単位と同一アイテムの組み合わせは1件 (409)
          - 同一アイテムの割当数量合計は所有数量以下 (422)
          - archive済みアイテム・archive済み収納単位へは割当できない (422)

        割当集合の競合を検知するため `expectedStorageUnitVersion` を要求し、
        成功時に収納単位のversionを1増加させる。
        responseは更新後の収納内容全体を返す。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageAllocationRequest"
      responses:
        "201":
          description: 割当に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitContentsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            version競合、または同一アイテムが既に本収納単位へ割当済みである
            (`STORAGE_ALLOCATION_ALREADY_EXISTS`)。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [storage-units]
      operationId: setStorageUnitAllocations
      summary: 収納単位の割当を一括置換する
      description: |
        設計書 12.4 の `setStorageUnitAllocations`。
        指定した割当集合へ置き換える。含まれない既存割当は削除する。

        `expectedStorageUnitVersion` で割当集合全体の競合を検知する。
        競合時はクライアント入力でserver状態を上書きせず 409 を返す。

        単一transactionで実行し、対象アイテム行を `SELECT FOR UPDATE` で
        ロックしてから割当数量合計を検証する (設計書 20章)。
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetStorageUnitAllocationsRequest"
      responses:
        "200":
          description: 置換に成功した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageUnitContentsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: less_session
      description: |
        session token Cookie。HttpOnly、SameSite=Lax、本番環境ではSecureを付与する。
        実際のCookie名は環境変数 `SESSION_COOKIE_NAME` で変更できる。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token。非HttpOnly Cookie `less_csrf` の値をそのまま送信する
        (signed double-submit cookie)。state変更requestで必須とする。
  parameters:
    PublicIdPathParameter:
      name: publicId
      in: path
      required: true
      description: 外部公開ID。内部IDはAPIへ公開しない (設計書 12.1)。
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ItemStorageAllocationListResponse:
      type: object
      description: |
        1アイテムの収納割当一覧。
        件数が収納単位数で上限づけられるためpaginationを持たない。
      required: [items, quantity, assignedQuantity, unassignedQuantity]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemStorageAllocationResponse"
        quantity:
          type: integer
          format: int32
          description: アイテムの所有数量
          example: 3
        assignedQuantity:
          type: integer
          format: int32
          example: 3
        unassignedQuantity:
          type: integer
          format: int32
          example: 0
...